Systematically format packages.
There was a loose standard to only use the `{package}` format the first time it was introduced, but then chapter 4 mixed that up. In the previous commits I systematically enforced that rule. Here I always format them. I like this way better, but I can roll back this commit if you prefer the other way.